+++ vkcs/blockstorage/import_volume_test.go
@@ -20,9 +20,10 @@ func TestAccBlockStorageVolume_importBasic(t *testing.T) {
 			},
 
 			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            resourceName,
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"metadata"},
 			},
 		},
 	})